+++ frontend/assets/js/genres.js
@@ -137,11 +137,28 @@ function deleteGenre(id) {
             "Content-Type": "application/json"
         }
 
-        const movieGenreUrl = `http://127.0.0.1:8085/movie_genre/movie/${id}`;
-        await fetch(movieGenreUrl, {
-            method: "DELETE",
+        const movieGenreUrl = `http://127.0.0.1:8085/movie_genre/genre/${id}`;
+
+        const responseMovie = await fetch(movieGenreUrl, {
+            method: "GET",
             headers: headersList
         });
+
+        if (responseMovie.ok) {
+            const relations = await responseMovie.json();
+            
+            if (relations.length > 0) {
+                await fetch(movieGenreUrl, {
+                    method: "DELETE",
+                    headers: headersList
+                });
+                console.log("Relaciones eliminadas.");
+            } else {
+                console.log("No había relaciones para borrar.");
+            }
+        } else {
+            console.log("No se pudo comprobar si había relaciones.");
+        }
 
         let response = await fetch(`http://127.0.0.1:8085/genres/${id}`, {
             method: "DELETE",
